worker: add RegistrationJob
This job type depends on a set of BuildJobs, and is meant to finalize
these image jobs once building and uploading has finished. The results
of the build jobs are passed to the registration jobs.

A typical example of this is a build job for AWS pushing the image into
S3, and then a registration job is triggered to import the image into
EC2. Having the registration job separate from the build job means that
the build worker can focus on building images, and registration jobs,
which mostly comes down to idling can all be one in parallel in a
dedicated registration worker.

Another example is building a set of images and uploading each into
Koji in a set of build jobs, and when they are all finished, CGImport
should be called from a registration job to finalize the transaction.
The call to CGImport cannot be done as part of one of the  build jobs as
results of all the image builds must be part of the call.

This functionality is not yet used, so this is not a functional change.

Signed-off-by: Tom Gundersen <teg@jklm.no>