WIP: upload to GCP
This patch is very much work-in-progress. It only attempts to upload
an image as a file, without actually importing it.

Credentials have to be retrieved from GCP and put into the container
config directory.

The current state of this patch is that it successfully loads the
credentials and the file to be uploaded, but then fails with
"Not Found".